refactor: replace to post header
