+++ src/lang/translations/ja/translation.json
@@ -136,6 +136,7 @@
   "characters": "文字",
   "saved": "保存しました",
   "add_localized_content": "ローカライズコンテンツの追加",
+  "add_to": "追加する",
   "dialog": {
     "confirm_deletion_title": "削除の確認",
     "confirm_deletion": "一度削除すると復元できません。削除してよろしいですか？",
